Move get_systems_within_jumps() to galaxy_topology.py
Move get_systems_within_jumps() to a new file galaxy_topology.py.
galaxy_topology.py will store common utility functions after starlanes
have been added to the galaxy   This anticipates and fixes cyclic import
problems with statistics.py, importing natives.py importing empires.py,
importing statistics.py
